Step two of simplifying UdapServer registration
